Project description
Alhamdulillah! I have successfully created a portfolio application  for my elder brother Galib Khan. The application consist of an admin panel through which the content can be managed very easily. The front end of this project is developed using Bootstrap and JavaScript. The backend section is developed using Laravel framework of PHP. Laravel framework is very interesting and time consuming. I will deployed it in real server within few days. I have produced the source code and database that may be helpful to the people who are interested this kind of project. The web site is designed carefully and it is responsive for mobile devices.

Designed with: HTML CSS Bootstrap
Frontend: JavaScript
Backend: Laravel
Database: MySQL